+++ react-ui/src/containers/ElementsNav/ElementsNav.js
@@ -3,25 +3,14 @@ import { connect } from 'react-redux';
 import propTypes from 'prop-types';
 import { ElementsNav as ElementsNavComponent } from '../../components';
 import s from '../../rootSelectors';
-// import { actions as sagaActions } from '../../sagas/actions';
 
 class ElementsNav extends Component {
-
-  onPreviousClick() {
-    return () => console.log('previous element');
-  }
-
-  onNextClick() {
-    return () => console.log('next element');
-  }
-
   render() {
     return this.props.elementsNavParams ?
       <ElementsNavComponent
         elementType={this.props.elementType}
+        completionPercentage={this.props.completionPercentage}
         {...this.props.elementsNavParams}
-        onPreviousClick={this.onPreviousClick.bind(this)}
-        onNextClick={this.onNextClick.bind(this)}
       />
       : null;
   }
@@ -29,12 +18,14 @@ class ElementsNav extends Component {
 
 ElementsNav.propTypes = {
   elementType: propTypes.string.isRequired,
-  elementsNavParams: propTypes.object
+  elementsNavParams: propTypes.object,
+  completionPercentage: propTypes.number // used to set 'completed' in Practice
 };
 
 const mapStateToProps = state => ({
   elementType: s.routing.getElementType(state),
-  elementsNavParams: s.getElementsNavParams(state)
+  elementsNavParams: s.getElementsNavParams(state),
+  completionPercentage: s.getProgressbarCompletion(state)
 });
 
 export default connect(